Update deployment target to iOS 8.4 and fix a bunch of compiler issues.

+++ libCronExpression/FieldFactory.m
@@ -51,7 +51,7 @@
 
     if(position >= [fields count])
     {
-        [NSException raise:@"Invalid argument" format:@"%d is not a valid position", position];
+        [NSException raise:@"Invalid argument" format:@"%lu is not a valid position", (unsigned long)position];
     }
     
     if([fields objectAtIndex: position] == [NSNull null])
